+++ gamewinner/bracket/parser.py
@@ -32,6 +32,13 @@ class Parser:
             self.west_plays = data["WestPlays"]
             self.year = data["Year"]
 
+            # Make an array of all teams arranged by region and by rank order.
+            # We can pop teams out one by one in each region to build the 4
+            # quadrants of the bracket. Order only matters in groups of 2, ie:
+            # if West plays East, the order doesn't matter as long as West and
+            # East are adjacent and South and Midwest are adjacent. This
+            # ensures that the winner of the West regional plays the winner of
+            # the `WestPlays`
             match self.west_plays:
                 case "East":
                     order = ["West", "East", "South", "Midwest"]
@@ -43,7 +50,7 @@ class Parser:
                     raise BracketLogicError("West can't play itself")
 
             for reg_str in order:
-                region = eval(f"self.{reg_str.lower()}")
+                region: dict[int, str] = eval(f"self.{reg_str.lower()}")
                 for rank in RANKS:
                     team = Team(
                         name=region.get(rank),
